Allow specifying a class for an element after which validation messages should appear.
When overriding an HTML select control with a jQuery UI combobox, the validation span will appear at the wrong location, being right between (the now invisible) select control and the span (by default with class name "custom-combobox") rendered by the jQuery UI combobox.

To fix this, it is possible to overrule the insertValidationMessage method to look for this specified span, but this is too hardcoded.

A second option is to provide the possibility to specify the class of the element after which the validation message should be rendered. This is the one I implemented here.